Esp-Terminal-App/requirements.txt
tkinter
pyserial